+++ src/TotalRecall.Core/Config.fs
@@ -65,6 +65,7 @@ type EmbeddingConfig = {
     BedrockModel: string option
     ModelName: string option
     ApiKey: string option
+    EmbeddingQueryPrefix: string option
 }
 
 type StorageConfig = {